# api/openapi.yaml
openapi: 3.0.3
info:
  title: HOA Management System API
  description: |
    REST API for the Sanderson Creek HOA Management System.

    ## Board Governance Features

    The board governance module provides:
    - Public/member access to current board roster (controlled by `board.visibility` feature flag)
    - Members-only access to historical board composition (controlled by `board.history-visibility` feature flag)
    - Public contact form with CAPTCHA and rate limiting
    - Admin CRUD operations for board titles and member assignments

    ## Democracy Module Features

    The democracy module provides transparent, hash-chained voting:
    - Poll creation and management (admin only)
    - Member voting with cryptographic receipt generation
    - Receipt verification for vote integrity
    - Optional member notifications via email
    - Support for informal, binding, and straw-poll types
    - Anonymous and identified voting modes

    ## Vendor Directory Features

    The vendor directory module provides curated vendor listings with moderation:
    - Public/member/admin tiered visibility with category filtering
    - Member submissions with admin moderation workflow
    - Rating and contact information managed by admins
    - Guest access limited to approved vendors in public categories
    - TTL-cached config for `vendors.public-categories` (60s)
    - Full audit trail for vendor CRUD and moderation actions

    ## Feature Flags

...

    ## Rate Limiting

...

...

...
  contact:
...
servers:
...
paths:
  /board:
    get:
      tags:
        - Board
      summary: Get current board roster
      description: |
        Returns the current board roster based on the `board.visibility` feature flag.

        - If `board.visibility` is `public`, returns roster to all users
        - If `board.visibility` is `members-only`, returns roster only to authenticated members
        - Unauthenticated users receive empty roster when visibility is members-only

        Results are sorted by board title rank (ascending).
      operationId: getCurrentRoster
      security:
        - {}
        - BearerAuth: []
      responses:
        '200':
          description: Board roster retrieved successfully
          headers:
            RateLimit-Limit:
              schema:
                type: integer
              description: Maximum requests per window
            RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            RateLimit-Reset:
              schema:
                type: integer
              description: Time when rate limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                type: object
                required:
                  - roster
                  - count
                properties:
                  roster:
                    type: array
                    items:
                      $ref: '#/components/schemas/RosterMember'
                  count:
                    type: integer
                    description: Number of roster members returned
              examples:
                public:
                  summary: Public roster response
                  value:
                    roster:
                      - id: 1
                        user:
                          id: 10
                          name: "Jane Smith"
                          email: "jane@example.com"
                        title:
                          id: 1
                          title: "President"
                          rank: 1
                        start_date: "2024-01-01"
                        bio: "Community leader with 5 years of HOA experience"
                        created_at: "2024-01-01T00:00:00.000Z"
                      - id: 2
                        user:
                          id: 15
                          name: "John Doe"
                          email: "john@example.com"
                        title:
                          id: 2
                          title: "Vice President"
                          rank: 2
                        start_date: "2024-01-01"
                        bio: null
                        created_at: "2024-01-01T00:00:00.000Z"
                    count: 2
                membersOnly:
                  summary: Empty response for guests when members-only
                  value:
                    roster: []
                    count: 0
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /polls:
    get:
      tags:
        - Polls
      summary: Get active and scheduled polls
      description: |
        Returns polls filtered by optional type and status parameters.

        **Access:**
        - Public and member access (no authentication required)
        - Optional authentication to adjust visibility of certain polls

        **Filtering:**
        - `type`: Filter by poll type (informal, binding, straw-poll)
        - `status`: Filter by status (active, scheduled, closed)

        **Feature Flags:**
        - Binding polls only appear if `polls.binding-enabled` is `true`

        **Status Determination:**
        - `scheduled`: Poll start time is in the future
        - `active`: Current time is between start_at and end_at
        - `closed`: Poll end time has passed
      operationId: getPolls
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by poll type
          schema:
            type: string
            enum: [informal, binding, straw-poll]
        - name: status
          in: query
          description: Filter by poll status
          schema:
            type: string
            enum: [active, scheduled, closed]
      responses:
        '200':
          description: Polls retrieved successfully
          headers:
            RateLimit-Limit:
              schema:
                type: integer
            RateLimit-Remaining:
              schema:
                type: integer
            RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                required:
                  - polls
                  - count
                  - serverTime
                properties:
                  polls:
                    type: array
                    items:
                      $ref: '#/components/schemas/PollSummary'
                  count:
                    type: integer
                  serverTime:
                    type: string
                    format: date-time
                    description: Server timestamp for countdown calculations
              examples:
                informal:
                  summary: Informal poll example
                  value:
                    polls:
                      - id: 1
                        title: "Should we add a new playground?"
                        description: "Community feedback on playground location"
                        type: "informal"
                        is_anonymous: false
                        start_at: "2025-01-15T00:00:00.000Z"
                        end_at: "2025-01-22T23:59:59.000Z"
                        status: "active"
                        options:
                          - id: 1
                            text: "Near the pool"
                            order_index: 0
                          - id: 2
                            text: "Near the tennis courts"
                            order_index: 1
                        created_by: "Admin User"
                        created_at: "2025-01-10T12:00:00.000Z"
                    count: 1
                    serverTime: "2025-01-18T14:30:00.000Z"
                binding:
                  summary: Binding poll example
                  value:
                    polls:
                      - id: 2
                        title: "Annual Budget Approval"
                        description: "Vote to approve the 2025 budget"
                        type: "binding"
                        is_anonymous: false
                        start_at: "2025-02-01T00:00:00.000Z"
                        end_at: "2025-02-07T23:59:59.000Z"
                        status: "scheduled"
                        options:
                          - id: 3
                            text: "Approve budget"
                            order_index: 0
                          - id: 4
                            text: "Reject budget"
                            order_index: 1
                        created_by: "Admin User"
                        created_at: "2025-01-15T09:00:00.000Z"
                    count: 1
                    serverTime: "2025-01-18T14:30:00.000Z"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Polls Admin
      summary: Create a new poll
      description: |
        Admin-only endpoint to create a new poll with options.

        **Validation:**
        - Title, type, start_at, end_at, and options (minimum 2) are required
        - End date must be after start date
        - Binding polls require `polls.binding-enabled` flag to be `true`

        **Workflow:**
        1. Validates input and feature flags
        2. Sanitizes title and description (strips HTML)
        3. Creates poll and poll options in transaction
        4. Logs audit event
        5. Optionally sends email notifications if `notify_members` is true

        **Feature Flags:**
        - `polls.binding-enabled`: Required for binding poll creation
        - `polls.notify-members-enabled`: Gates email notification functionality
      operationId: createPoll
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollCreateRequest'
            examples:
              informal:
                summary: Informal poll creation
                value:
                  title: "Should we add a new playground?"
                  description: "Community feedback on playground location"
                  type: "informal"
                  is_anonymous: false
                  notify_members: true
                  start_at: "2025-01-15T00:00:00.000Z"
                  end_at: "2025-01-22T23:59:59.000Z"
                  options:
                    - text: "Near the pool"
                      order_index: 0
                    - text: "Near the tennis courts"
                      order_index: 1
                    - text: "No new playground"
                      order_index: 2
              binding:
                summary: Binding poll creation
                value:
                  title: "Annual Budget Approval"
                  description: "Vote to approve the 2025 budget of $150,000"
                  type: "binding"
                  is_anonymous: false
                  notify_members: true
                  start_at: "2025-02-01T00:00:00.000Z"
                  end_at: "2025-02-07T23:59:59.000Z"
                  options:
                    - text: "Approve budget"
                      order_index: 0
                    - text: "Reject budget"
                      order_index: 1
      responses:
        '201':
          description: Poll created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - poll
                properties:
                  message:
                    type: string
                  poll:
                    $ref: '#/components/schemas/PollDetail'
              example:
                message: "Poll created successfully"
                poll:
                  id: 1
                  title: "Should we add a new playground?"
                  description: "Community feedback on playground location"
                  type: "informal"
                  is_anonymous: false
                  notify_members: true
                  start_at: "2025-01-15T00:00:00.000Z"
                  end_at: "2025-01-22T23:59:59.000Z"
                  created_by: 1
                  created_at: "2025-01-10T12:00:00.000Z"
                  options:
                    - id: 1
                      poll_id: 1
                      text: "Near the pool"
                      order_index: 0
                    - id: 2
                      poll_id: 1
                      text: "Near the tennis courts"
                      order_index: 1
        '400':
          description: Bad request - Missing fields or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    message: "Missing required fields: title, type, start_at, end_at, options"
                invalidDates:
                  summary: Invalid date range
                  value:
                    message: "End date must be after start date"
                invalidType:
                  summary: Invalid poll type
                  value:
                    message: "Invalid poll type. Must be one of: informal, binding, straw-poll"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Binding polls are currently disabled"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /vendors:
    get:
      tags:
        - Vendors
      summary: Get vendors with optional filtering
      description: |
        Returns vendors based on user authentication status and role.

        **Guest Users (unauthenticated):**
        - See only vendors in categories listed in `vendors.public-categories` config flag
        - See only approved vendors with public visibility
        - See only name and category (no contact info or ratings)

        **Authenticated Members:**
        - See approved vendors with public or member visibility
        - See name, category, contact info, and ratings

        **Admins:**
        - See all vendors regardless of moderation state or visibility
        - See all fields including moderation state and notes

        Results are sorted by category (ascending), then name (ascending).
      operationId: getVendors
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by service category
          schema:
            type: string
          example: "Landscaping"
        - name: search
          in: query
          description: Search vendors by name (partial match)
          schema:
            type: string
          example: "Green"
        - name: status
          in: query
          description: Filter by moderation state (admin only)
          schema:
            type: string
            enum: [pending, approved, denied]
          example: "pending"
      responses:
        '200':
          description: Vendors retrieved successfully
          headers:
            RateLimit-Limit:
              schema:
                type: integer
            RateLimit-Remaining:
              schema:
                type: integer
            RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                required:
                  - vendors
                  - count
                properties:
                  vendors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vendor'
                  count:
                    type: integer
                    description: Number of vendors returned
                  filters:
                    type: object
                    properties:
                      category:
                        type: string
                        nullable: true
                      search:
                        type: string
                        nullable: true
                      status:
                        type: string
                        nullable: true
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Vendors
      summary: Create a new vendor
      description: |
        Submit a new vendor entry.

        **Members:** Vendor submitted for moderation (returns HTTP 202 Accepted)
        **Admins:** Vendor directly approved (returns HTTP 201 Created)

        Rate-limited to prevent abuse.
      operationId: createVendor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - service_category
              properties:
                name:
                  type: string
                  description: Vendor business name
                  example: "Green Thumb Landscaping"
                service_category:
                  type: string
                  description: Service category
                  example: "Landscaping"
                contact_info:
                  type: string
                  nullable: true
                  description: Contact information (phone, email, address)
                  example: "555-1234, info@greenthumb.com"
                rating:
                  type: integer
                  nullable: true
                  minimum: 1
                  maximum: 5
                  description: Internal rating (1-5 scale)
                  example: 4
                notes:
                  type: string
                  nullable: true
                  description: Internal notes about vendor
                  example: "Reliable service, good pricing"
                visibility_scope:
                  type: string
                  enum: [public, members, admins]
                  description: Visibility level
                  example: "members"
      responses:
        '201':
          description: Vendor created successfully (admin)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vendor created successfully"
                  vendor:
                    $ref: '#/components/schemas/VendorFull'
        '202':
          description: Vendor submitted for moderation (member)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vendor submitted for moderation"
                  vendor:
                    $ref: '#/components/schemas/VendorFull'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoints.
        Admin role required for administrative operations.
  schemas:
    RosterMember:
      type: object
      required:
        - id
        - user
        - title
        - start_date
        - created_at
      properties:
        id:
          type: integer
          description: Board member record ID
        user:
          type: object
          required:
            - id
            - name
            - email
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
              format: email
        title:
          type: object
          required:
            - id
            - title
            - rank
          properties:
            id:
              type: integer
            title:
              type: string
            rank:
              type: integer
        start_date:
          type: string
          format: date
          description: Date the member assumed this position
        bio:
          type: string
          nullable: true
          description: Optional biography or description
        created_at:
          type: string
          format: date-time
          description: Timestamp when record was created
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error message
    PollSummary:
      type: object
      required:
        - id
        - title
        - type
        - is_anonymous
        - start_at
        - end_at
        - status
        - options
        - created_by
        - created_at
      properties:
        id:
          type: integer
          description: Poll ID
        title:
          type: string
          description: Poll title
        description:
          type: string
          nullable: true
          description: Optional poll description
        type:
          type: string
          enum: [informal, binding, straw-poll]
          description: Poll type
        is_anonymous:
          type: boolean
          description: Whether votes are anonymous
        start_at:
          type: string
          format: date-time
          description: Poll start timestamp
        end_at:
          type: string
          format: date-time
          description: Poll end timestamp
        status:
          type: string
          enum: [scheduled, active, closed]
          description: Poll status calculated server-side
        options:
          type: array
          items:
            $ref: '#/components/schemas/PollOptionSummary'
          description: Poll options
        created_by:
          type: string
          description: Name of poll creator
        created_at:
          type: string
          format: date-time
          description: Poll creation timestamp
    PollDetail:
      type: object
      required:
        - id
        - title
        - type
        - is_anonymous
        - start_at
        - end_at
        - status
        - options
        - created_by
        - created_at
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [informal, binding, straw-poll]
        is_anonymous:
          type: boolean
        notify_members:
          type: boolean
          description: Whether members were notified
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, active, closed]
        options:
          type: array
          items:
            $ref: '#/components/schemas/PollOptionSummary'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: '#/components/schemas/PollResult'
          nullable: true
          description: Vote counts (only for closed polls when requested)
    PollOptionSummary:
      type: object
      required:
        - id
        - text
        - order_index
      properties:
        id:
          type: integer
          description: Option ID
        text:
          type: string
          description: Option text
        order_index:
          type: integer
          description: Display order
    PollCreateRequest:
      type: object
      required:
        - title
        - type
        - start_at
        - end_at
        - options
      properties:
        title:
          type: string
          description: Poll title
        description:
          type: string
          nullable: true
          description: Optional poll description
        type:
          type: string
          enum: [informal, binding, straw-poll]
          description: Poll type
        is_anonymous:
          type: boolean
          default: false
          description: Whether votes should be anonymous
        notify_members:
          type: boolean
          default: false
          description: Send email notifications to members
        start_at:
          type: string
          format: date-time
          description: Poll start timestamp
        end_at:
          type: string
          format: date-time
          description: Poll end timestamp
        options:
          type: array
          minItems: 2
          items:
            type: object
            required:
              - text
            properties:
              text:
                type: string
                description: Option text
              order_index:
                type: integer
                description: Optional display order (defaults to array index)
...
    PollResult:
      type: object
      required:
        - option_id
        - text
        - vote_count
      properties:
        option_id:
          type: integer
          description: Option ID
        text:
          type: string
          description: Option text
        vote_count:
          type: integer
          description: Number of votes for this option
    Vendor:
      type: object
      description: Vendor DTO with fields visible based on user role
      required:
        - id
        - name
        - service_category
        - visibility_scope
      properties:
        id:
          type: integer
          description: Vendor ID
        name:
          type: string
          description: Vendor business name
        service_category:
          type: string
          description: Service category
        visibility_scope:
          type: string
          enum: [public, members, admins]
          description: Visibility level
        contact_info:
          type: string
          nullable: true
          description: Contact information (visible to members and admins only)
        rating:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          description: Internal rating (visible to members and admins only)
...
  responses:
    Unauthorized:
      description: Unauthorized - Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Unauthorized"
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      headers:
        RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests per window
        RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window (will be 0)
        RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Too many requests from this IP, please try again later."
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Internal server error"